Finish up listen docs
