+++ src/listener/mod.rs
@@ -3,8 +3,10 @@
 use async_trait::async_trait;
 
 mod concurrent_listener;
-mod connection_info;
 mod failover_listener;
+mod listen_info;
+mod report;
+
 #[cfg(feature = "h1-server")]
 mod parsed_listener;
 #[cfg(feature = "h1-server")]
@@ -15,17 +17,15 @@ mod to_listener_impls;
 #[cfg(all(unix, feature = "h1-server"))]
 mod unix_listener;
 
-use std::{
-    fmt::{Debug, Display},
-    future::Future,
-};
+use std::fmt::{Debug, Display};
 
 use crate::Server;
 use async_std::io;
 
 pub use concurrent_listener::ConcurrentListener;
-pub use connection_info::ConnectionInfo;
 pub use failover_listener::FailoverListener;
+pub use listen_info::ListenInfo;
+pub use report::Report;
 pub use to_listener::ToListener;
 
 #[cfg(feature = "h1-server")]
@@ -49,29 +49,6 @@ where
     /// is called exactly once, and is expected to spawn tasks for
     /// each incoming connection.
     async fn listen_with(&mut self, app: Server<State>, f: F) -> io::Result<()>;
-
-    // /// This is a shorthand for `listen_with` with an empty closure.
-    // async fn listen(&mut self, app: Server<State>) -> io::Result<()> {
-    //     self.listen_with(app, |server| async move { Ok(server) })
-    //         .await
-    // }
-}
-
-#[async_trait]
-pub trait Report: Clone + Send + Sync + 'static {
-    async fn call(&self, server: ConnectionInfo) -> io::Result<()>;
-}
-
-#[async_trait]
-impl<F, Fut> Report for F
-where
-    F: Clone + Send + Sync + 'static + Fn(ConnectionInfo) -> Fut,
-    Fut: Future<Output = io::Result<()>> + Send + 'static,
-{
-    async fn call(&self, info: ConnectionInfo) -> io::Result<()> {
-        let fut = (self)(info);
-        Ok(fut.await?)
-    }
 }
 
 /// crate-internal shared logic used by tcp and unix listeners to
