+++ src/listener/tcp_listener.rs
@@ -1,4 +1,4 @@
-use super::{is_transient_error, ConnectionInfo};
+use super::{is_transient_error, ListenInfo};
 
 use crate::listener::Listener;
 use crate::{log, Server};
@@ -77,7 +77,7 @@ where
     async fn listen_with(&mut self, app: Server<State>, f: F) -> io::Result<()> {
         self.connect().await?;
         let listener = self.listener()?;
-        let info = ConnectionInfo::new(format!("{}", self));
+        let info = ListenInfo::new(format!("{}", self));
         f.call(info).await?;
 
         let mut incoming = listener.incoming();
